+++ src/professions/profession-versions.service.spec.ts
@@ -10,6 +10,7 @@ import {
   SelectQueryBuilder,
 } from 'typeorm';
 import { ProfessionVersionsService } from './profession-versions.service';
+import { ProfessionsSearchService } from './professions-search.service';
 import professionVersionFactory from '../testutils/factories/profession-version';
 import {
   ProfessionVersion,
@@ -25,9 +26,11 @@ describe('ProfessionVersionsService', () => {
   let service: ProfessionVersionsService;
   let repo: Repository<ProfessionVersion>;
   let connection: DeepMocked<Connection>;
+  let searchService: DeepMocked<ProfessionsSearchService>;
 
   beforeEach(async () => {
     connection = createMock<Connection>();
+    searchService = createMock<ProfessionsSearchService>();
 
     const module: TestingModule = await Test.createTestingModule({
       providers: [
@@ -39,6 +42,10 @@ describe('ProfessionVersionsService', () => {
         {
           provide: Connection,
           useValue: connection,
+        },
+        {
+          provide: ProfessionsSearchService,
+          useValue: searchService,
         },
       ],
     }).compile();
@@ -210,6 +217,11 @@ describe('ProfessionVersionsService', () => {
         status: ProfessionVersionStatus.Live,
       });
 
+      expect(searchService.index).toBeCalledWith({
+        ...version,
+        status: ProfessionVersionStatus.Live,
+      });
+
       expect(queryRunner.commitTransaction).toHaveBeenCalled();
       expect(queryRunner.release).toHaveBeenCalled();
     });
@@ -243,6 +255,11 @@ describe('ProfessionVersionsService', () => {
       });
 
       expect(saveSpy).toHaveBeenCalledWith({
+        ...liveVersion,
+        status: ProfessionVersionStatus.Archived,
+      });
+
+      expect(searchService.delete).toHaveBeenCalledWith({
         ...liveVersion,
         status: ProfessionVersionStatus.Archived,
       });