+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Index profession versions in Opensearch when publishing
+- Delete previously indexed profession versions from Opensearch when publishing
+
 ## [release-009] - 2022-03-11
 
 ### Changed